+++ bindings_ffi/src/lib.rs
@@ -23,7 +23,7 @@ pub enum GenericError {
     #[error("Storage error: {0}")]
     Storage(#[from] xmtp_mls::storage::StorageError),
     #[error("API error: {0}")]
-    ApiError(#[from] xmtp_proto::api_client::Error),
+    ApiError(#[from] xmtp_proto::Error),
     #[error("Group error: {0}")]
     GroupError(#[from] xmtp_mls::groups::GroupError),
     #[error("Signature: {0}")]
