+++ xmtp_proto/src/api_client.rs
@@ -1,7 +1,5 @@
-use std::{error::Error as StdError, fmt};
-
 use futures::Stream;
-use crate::api_client::ErrorKind::DecodingError;
+use crate::Error;
 pub use super::xmtp::message_api::v1::{
     BatchQueryRequest, BatchQueryResponse, Envelope, PagingInfo, PublishRequest, PublishResponse,
     QueryRequest, QueryResponse, SubscribeRequest,
@@ -96,95 +94,6 @@ pub mod trait_impls {
             T: XmtpMlsClient + XmtpIdentityClient + ClientWithMetadata + Send + Sync + ?Sized
         {
         }
-    }
-}
-
-#[derive(Debug)]
-pub enum ErrorKind {
-    SetupCreateChannelError,
-    SetupTLSConfigError,
-    SetupConnectionError,
-    PublishError,
-    QueryError,
-    SubscribeError,
-    BatchQueryError,
-    MlsError,
-    IdentityError,
-    SubscriptionUpdateError,
-    MetadataError,
-    MissingPayloadError,
-    DecodingError(hex::FromHexError),
-}
-
-type ErrorSource = Box<dyn StdError + Send + Sync + 'static>;
-
-pub struct Error {
-    kind: ErrorKind,
-    source: Option<ErrorSource>,
-}
-
-impl Error {
-    pub fn new(kind: ErrorKind) -> Self {
-        Self { kind, source: None }
-    }
-
-    pub fn with(mut self, source: impl Into<ErrorSource>) -> Self {
-        self.source = Some(source.into());
-        self
-    }
-}
-
-impl From<hex::FromHexError> for Error {
-    fn from(err: hex::FromHexError) -> Self {
-        Error::new(DecodingError(err))
-    }
-}
-
-impl fmt::Debug for Error {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let mut f = f.debug_tuple("xmtp::error::Error");
-
-        f.field(&self.kind);
-
-        if let Some(source) = &self.source {
-            f.field(source);
-        }
-
-        f.finish()
-    }
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let s = match &self.kind {
-            ErrorKind::SetupCreateChannelError => "failed to create channel",
-            ErrorKind::SetupTLSConfigError => "tls configuration failed",
-            ErrorKind::SetupConnectionError => "connection failed",
-            ErrorKind::PublishError => "publish error",
-            ErrorKind::QueryError => "query error",
-            ErrorKind::SubscribeError => "subscribe error",
-            ErrorKind::BatchQueryError => "batch query error",
-            ErrorKind::IdentityError => "identity error",
-            ErrorKind::MlsError => "mls error",
-            ErrorKind::SubscriptionUpdateError => "subscription update error",
-            ErrorKind::MetadataError => "metadata error",
-            ErrorKind::MissingPayloadError => "missing payload error",
-            DecodingError(_) => "could not convert payload from hex"
-        };
-        f.write_str(s)?;
-        if self.source().is_some() {
-            f.write_str(": ")?;
-            f.write_str(&self.source().unwrap().to_string())?;
-        }
-        Ok(())
-    }
-}
-
-impl StdError for Error {
-    fn source(&self) -> Option<&(dyn StdError + 'static)> {
-        self.source
-            .as_ref()
-            .map(|source| &**source as &(dyn StdError + 'static))
     }
 }
