+++ xmtp_proto/src/convert.rs
@@ -10,10 +10,9 @@ use crate::xmtp::xmtpv4::envelopes::{
 };
 use crate::xmtp::xmtpv4::payer_api::PublishClientEnvelopesRequest;
 
-use crate::api_client;
-use crate::api_client::Error;
-use crate::api_client::ErrorKind::MissingPayloadError;
-
+use crate::{Error};
+use crate::ErrorKind::InternalError;
+use crate::InternalError::{MissingPayloadError};
 use crate::v4_utils::{build_identity_topic_from_hex_encoded, build_welcome_message_topic, get_group_message_topic, get_key_package_topic};
 
 mod inbox_id {
@@ -35,7 +34,7 @@ mod inbox_id {
 }
 
 impl TryFrom<UploadKeyPackageRequest> for PublishClientEnvelopesRequest {
-    type Error = api_client::Error;
+    type Error = Error;
 
     fn try_from(req: UploadKeyPackageRequest) -> Result<Self, Error> {
         if let Some(key_package) = req.key_package.as_ref() {
@@ -48,13 +47,13 @@ impl TryFrom<UploadKeyPackageRequest> for PublishClientEnvelopesRequest {
                 }],
             })
         } else {
-            Err(Error::new(MissingPayloadError))
+            Err(Error::new(InternalError(MissingPayloadError)))
         }
     }
 }
 
 impl TryFrom<PublishIdentityUpdateRequest> for PublishClientEnvelopesRequest {
-    type Error = api_client::Error;
+    type Error = Error;
 
     fn try_from(req: PublishIdentityUpdateRequest) -> Result<Self, Error> {
         if let Some(identity_update) = req.identity_update {
@@ -67,13 +66,13 @@ impl TryFrom<PublishIdentityUpdateRequest> for PublishClientEnvelopesRequest {
                 }],
             })
         } else {
-            Err(Error::new(MissingPayloadError))
+            Err(Error::new(InternalError(MissingPayloadError)))
         }
     }
 }
 
 impl TryFrom<GroupMessageInput> for PublishClientEnvelopesRequest {
-    type Error = api_client::Error;
+    type Error = crate::Error;
 
     fn try_from(req: GroupMessageInput) -> Result<Self, Error> {
         if let Some(GroupMessageInputVersion::V1(ref version)) = req.version {
@@ -86,13 +85,13 @@ impl TryFrom<GroupMessageInput> for PublishClientEnvelopesRequest {
                 }],
             })
         } else {
-            Err(Error::new(MissingPayloadError))
+            Err(Error::new(InternalError(MissingPayloadError)))
         }
     }
 }
 
 impl TryFrom<WelcomeMessageInput> for PublishClientEnvelopesRequest {
-    type Error = api_client::Error;
+    type Error = crate::Error;
 
     fn try_from(req: WelcomeMessageInput) -> Result<Self, Self::Error> {
         if let Some(WelcomeMessageVersion::V1(ref version)) = req.version {
@@ -105,7 +104,7 @@ impl TryFrom<WelcomeMessageInput> for PublishClientEnvelopesRequest {
                 }],
             })
         } else {
-            Err(Error::new(MissingPayloadError))
+            Err(Error::new(InternalError(MissingPayloadError)))
         }
     }
 }
