+++ xmtp_proto/src/lib.rs
@@ -4,6 +4,9 @@ mod generated {
 }
 pub use generated::*;
 
+mod error;
+pub use error::*;
+
 #[cfg(feature = "xmtp-message_api-v1")]
 pub mod api_client;
